\docType{class}
\name{ExonCountSet-class}
\alias{ExonCountSet}
\alias{ExonCountSet-class}
\title{
  This class has been deprecated and will become defunct.
}
\description{
  The ExonCountSet object has been deprecated and substituted by
  the DEXSeqDataSet object.  Therefore, all the functions and methods
  for the ExonCountSet object will be deprecated and defunct at the 
  next releases. These functions have been subsituted by new functions 
  that are compatible with the newer versions of DEXSeq.
}
