\name{DEUresultTable}
\alias{DEUresultTable}
\title{
      Makes a result table from \code{testForDEU}.
}
\description{
      This function makes a table describing the dispersion parameter used in each of the test and the results from \code{testForDEU}.
}
\usage{
DEUresultTable(ecs)
}
\arguments{
  \item{ecs}{
     An ExonCountSet object.
	}
}
\value{
	A data frame that summarizes the results from \code{testForDEU}.
}
