\name{DEXSeq-defunct}
\alias{DEXSeq-defunct}
\alias{prepareAnnotationForDEXSeq}
\alias{prepareAnnotationForDEXSeq-defunct}
\alias{countReadsForDEXSeq}
\alias{countReadsForDEXSeq-defunct}
\title{
  This functions are defunct and will be deleted
}
\description{
  The functions that are defunct, they will be removed
  in the next release.
}
